% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_tagger.R
\name{make_target_tagger_function}
\alias{make_target_tagger_function}
\title{Make tagger function for given list of target points}
\usage{
make_target_tagger_function(targets_sf, id_column = "target_id")
}
\arguments{
\item{targets_sf}{An sf data frame of target points}

\item{id_column}{The name of the id column}
}
\value{
A function that takes a GPX track as an argument
}
\description{
Given an sf data frame with a list of targets, this will create a
function that determines which a given track intersects. The sf data frame
must have an id_column and the required sf "geometry" column and should be in
the GPS (4326) coordinate reference system.
}
\examples{
a_target <- sf::st_as_sf(
    data.frame(target_id="a_view", lat=42.17, lon=-74.35),
    coords=2:3, crs=4326)
target_tagger <- make_target_tagger_function(a_target)
}
